# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Muttley API
  description: >-
    API REST do Sistema Muttley para Gestão de Eventos Acadêmicos, Inscrições,
    Emissão e Validação de Certificados com Assinatura Visual e Gamificação por Medalhas.
    Desenvolvido para a FATEC Zona Leste.
  version: 0.1.0-alpha
  contact:
    name: Equipe Muttley - FATEC Zona Leste
    url: https://github.com/andrelamego/Backend-Muttley
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8083
    description: Servidor Local de Desenvolvimento
...
paths:
...
  /api/auth/register:
    post:
      tags:
        - Autenticação
      summary: Solicitar cadastro
      description: Cria uma conta USER pendente e envia um convite de uso único para definir a senha.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SolicitacaoCadastro'
      responses:
        '202':
          description: Solicitação recebida
        '400':
          description: Dados inválidos
    put:
      tags:
        - Autenticação
      summary: Completar cadastro
      description: Define a senha mediante convite válido de uso único recebido por email.
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizacaoPessoa'
      responses:
        '200':
          description: Cadastro completado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
        '404':
          description: Convite inválido ou expirado
  /api/auth/login:
    post:
      tags:
        - Autenticação
      summary: Autenticar usuário (Login)
      description: Valida credenciais (email e senha) e retorna o token JWT de acesso.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Autenticado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Email ou senha inválidos
...
  /api/certificados/{codigo}:
    get:
      tags:
        - Certificados Públicos
      summary: Consultar dados públicos do certificado
      parameters:
        - name: codigo
          in: path
          required: true
          schema:
            type: string
            example: 'ABC123XYZ'
      responses:
        '200':
          description: Certificado encontrado
        '404':
          description: Certificado não encontrado
  /api/certificados/{codigo}/preview:
    get:
      tags:
        - Certificados Públicos
      summary: Pré-visualizar PDF do certificado
      parameters:
        - name: codigo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PDF gerado
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          description: Não encontrado
  /api/certificados/{codigo}/download:
    get:
      tags:
        - Certificados Públicos
      summary: Baixar PDF do certificado
      parameters:
        - name: codigo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Download iniciado
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          description: Não encontrado
  /{id}/assinatura-visual:
    get:
      tags:
        - Certificados Públicos
      summary: Exibir imagem pública da assinatura
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Imagem da assinatura
        '404':
          description: Não encontrada
...
  /api/admin/certificados/{id}/assinatura-visual:
    get:
      tags:
        - Administração - Certificados
      summary: Exibir imagem da assinatura visual do certificado
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Imagem retornada
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Insira o token JWT retornado no endpoint /api/auth/login.
  schemas:
    SolicitacaoCadastro:
      type: object
      required: [nome, email]
      properties:
        nome:
          type: string
        email:
          type: string
          format: email
    LoginRequest:
      type: object
      required:
        - email
        - senha
      properties:
        email:
          type: string
          format: email
          example: usuario@fatec.sp.gov.br
        senha:
          type: string
          example: Senha@1234
    UsuarioResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nome:
          type: string
          example: João da Silva
        email:
          type: string
          example: joao@fatec.sp.gov.br
        role:
          type: string
          enum: [USER, ADMIN]
          example: USER
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
          example: 7200
        usuario:
          $ref: '#/components/schemas/UsuarioResponse'
    AtualizacaoPessoa:
      type: object
      required: [nome, email, telefone, cpf, senha]
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
          example: Maria Souza
        email:
          type: string
          format: email
          example: maria@fatec.sp.gov.br
        cpf:
          type: string
          example: '529.982.247-25'
        telefone:
          type: string
          example: '11999999999'
        senha:
          type: string
          example: SenhaForte@123
...